Add files via upload
In this assignment the model is used to predict bounding boxes for some birds which will be compared with reference boxes of the same bird images. Model efficiency will be computed based on the proportion of images that achieved an iou score greater or equal to 0.5 with the reference boxes.  At least a 45% proportion was required to pass this assignment.